feat:implemented rifgts-page
