+++ src/locales/ua.json
@@ -73,6 +73,47 @@
     },
     "favorites": "Вибране",
     "search": "Пошук продукту",
-    "back": "Назад"
+    "back": "Назад",
+    "rights": {
+      "title": "Права",
+      "titleInfo": "Право на інформацію",
+      "titleReturn": "Право на повернення",
+      "titleConfidentiality": "Право на конфіденційність",
+      "textInfo": {
+        "part1": "Клієнти мають право отримувати ",
+        "highlight1": "точну",
+        "part2": " та ",
+        "highlight2": "вичерпну інформацію",
+        "part3": " про продукти та послуги, які пропонує наша компанія. Це включає деталі про ",
+        "highlight3": "характеристики продукту",
+        "highlight4": "використання",
+        "highlight5": "гарантії",
+        "part4": " та будь-які пов'язані ",
+        "highlight6": "умови та положення",
+        "part5": "Прозорість є нашим пріоритетом."
+      },
+      "textReturn": {
+        "part1": "Покупці мають право повернути продукти відповідно до умов, зазначених у нашій ",
+        "highlight1": "політиці повернення",
+        "part2": ". Це включає збереження ",
+        "highlight2": "оригінального стану продукту",
+        "part3": " та ",
+        "highlight3": "цілісності упаковки",
+        "part4": ". Будь ласка, ознайомтеся з нашими детальними ",
+        "highlight4": "рекомендаціями щодо повернення",
+        "part5": ", щоб забезпечити простий і безпроблемний процес."
+      },
+      "textConfidentiality": {
+        "part1": "Ми прагнемо забезпечити ",
+        "highlight1": "безпеку",
+        "part2": " та ",
+        "highlight2": "конфіденційність",
+        "part3": " ваших особистих даних. Вся інформація, надана клієнтами, обробляється і зберігається строго відповідно до чинного ",
+        "highlight3": "законодавства про захист даних",
+        "part4": " та нормативних актів. Ваша ",
+        "highlight4": "приватність",
+        "part5": " є важливою для нас."
+      }
+    }
   }
 }
