+++ src/styles/abstract/_constants.scss
@@ -76,6 +76,7 @@ $transition-time: 0.3s;
 $tablet-l-min-whidth: 772px;
 $tablet-xl-min-whidth: 930px;
 $tablet-min-whidth: 640px;
+$mobile-s-width: 380px;
 $mobile-l-width: 550px;
 $laptop-min-whidth: 1024px;
 $desctop-min-whidth: 1200px;
